requirements.txt
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.42
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
SQLAlchemy==2.0.40
urllib3==2.4.0
